Fixed more address types related issues. Mostly printf() or comparison warnings, but also some oversights from earlier changes.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@37000 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/system/kernel/cache/vnode_store.cpp
@@ -52,7 +52,7 @@ VMVnodeCache::Read(off_t offset, const generic_io_vec *vecs, size_t count,
 
 	generic_size_t bytesEnd = *_numBytes;
 
-	if (offset + bytesEnd > virtual_end)
+	if (offset + (off_t)bytesEnd > virtual_end)
 		bytesEnd = virtual_end - offset;
 
 	// If the request could be filled completely, or an error occured,
